+++ src/main/kotlin/com/autonomousapps/internal/moshi.kt
@@ -15,47 +15,67 @@ val MOSHI: Moshi by lazy {
         .build()
 }
 
-inline fun <reified T> getJsonAdapter(): JsonAdapter<T> {
-    return MOSHI.adapter(T::class.java)
+inline fun <reified T> getJsonAdapter(withNulls: Boolean = false): JsonAdapter<T> {
+    val adapter = MOSHI.adapter(T::class.java)
+    return if (withNulls) {
+        adapter.serializeNulls()
+    } else {
+        adapter
+    }
 }
 
-inline fun <reified T> getJsonListAdapter(): JsonAdapter<List<T>> {
+inline fun <reified T> getJsonListAdapter(withNulls: Boolean = false): JsonAdapter<List<T>> {
     val type = newParameterizedType(List::class.java, T::class.java)
-    return MOSHI.adapter(type)
+    val adapter = MOSHI.adapter<List<T>>(type)
+    return if (withNulls) {
+        adapter.serializeNulls()
+    } else {
+        adapter
+    }
 }
 
-inline fun <reified T> getJsonSetAdapter(): JsonAdapter<Set<T>> {
+inline fun <reified T> getJsonSetAdapter(withNulls: Boolean = false): JsonAdapter<Set<T>> {
     val type = newParameterizedType(Set::class.java, T::class.java)
-    return MOSHI.adapter(type)
+    val adapter = MOSHI.adapter<Set<T>>(type)
+    return if (withNulls) {
+        adapter.serializeNulls()
+    } else {
+        adapter
+    }
 }
 
-inline fun <reified K, reified V> getJsonMapAdapter(): JsonAdapter<Map<K, V>> {
+inline fun <reified K, reified V> getJsonMapAdapter(withNulls: Boolean = false): JsonAdapter<Map<K, V>> {
     val type = newParameterizedType(Map::class.java, K::class.java, V::class.java)
-    return MOSHI.adapter(type)
+    val adapter = MOSHI.adapter<Map<K, V>>(type)
+    return if (withNulls) {
+        adapter.serializeNulls()
+    } else {
+        adapter
+    }
 }
 
 inline fun <reified T> String.fromJson(): T {
     return getJsonAdapter<T>().fromJson(this)!!
 }
 
-inline fun <reified T> T.toJson(): String {
-    return getJsonAdapter<T>().toJson(this)
+inline fun <reified T> T.toJson(withNulls: Boolean = false): String {
+    return getJsonAdapter<T>(withNulls).toJson(this)
 }
 
 inline fun <reified T> String.fromJsonList(): List<T> {
     return getJsonListAdapter<T>().fromJson(this)!!
 }
 
-inline fun <reified T> List<T>.toPrettyString(): String {
-    return getJsonListAdapter<T>().indent("  ").toJson(this)
+inline fun <reified T> List<T>.toPrettyString(withNulls: Boolean = false): String {
+    return getJsonListAdapter<T>(withNulls).indent("  ").toJson(this)
 }
 
-inline fun <reified T> Set<T>.toPrettyString(): String {
-    return getJsonSetAdapter<T>().indent("  ").toJson(this)
+inline fun <reified T> Set<T>.toPrettyString(withNulls: Boolean = false): String {
+    return getJsonSetAdapter<T>(withNulls).indent("  ").toJson(this)
 }
 
-inline fun <reified K, reified V> Map<K, V>.toPrettyString(): String {
-    return getJsonMapAdapter<K, V>().indent("  ").toJson(this)
+inline fun <reified K, reified V> Map<K, V>.toPrettyString(withNulls: Boolean = false): String {
+    return getJsonMapAdapter<K, V>(withNulls).indent("  ").toJson(this)
 }
 
 @Suppress("unused", "HasPlatformType")